requirements.txt
aiohttp==3.8.1
aiosignal==1.3.1
altgraph==0.17.3
async-timeout==4.0.2
attrs==23.1.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==40.0.2
customtkinter==5.1.2
darkdetect==0.8.0
Deprecated==1.2.14
frozenlist==1.3.3
idna==3.4
multidict==6.0.4
pefile==2023.2.7
Pillow==9.5.0
pycparser==2.21
pyinstaller==5.11.0
pyinstaller-hooks-contrib==2023.3
PyJWT==2.7.0
PyNaCl==1.5.0
pywin32-ctypes==0.2.0
requests==2.31.0
typing_extensions==4.6.2
urllib3==2.0.2
wrapt==1.15.0
yarl==1.9.2
